In our array of canvases, store the selections as well.  Now, those can be transformed from here.
In the update() function, reduce code forking, by storing the value of
h, and using a lookup dictionary for the y transforms.  Perform
transformations on the canvas selections, and pass in the new height()
and call the update() API functions on the canvas objects.  This now
does the combine/separate transitions properly.