Add a new 'domain' private var to store the x and y domains for the dataset.  This should be able to get passed in (as the dataset has this information already) or calculated.
The scale.x and scale.y domains are now set by the domain var.

The .zoom() API function now transitions to the new zoom extent.  A new
API function, .snap(), snaps to the new zoom extent.  Not sure these
necessarily need to be separate functions, but that can be optimized
later.