Allow local installation of gems with bundle
This commit allows local installation of ruby gems with `bundle install --path vendor/bundle`.  `.gitignore` rules are added to prevent the `vendor/bundle` and `.bundle` directories being checked in, and the `vendor` path prefix is added to the `exclude` key in `_config.yaml` so that Jekyll doesn't try to build everything in there.