Reapply layout
Quickly re-establish the previous layout, fixing a few layout bugs at the same time.  Looks like we'll be hacking this layout to pieces anyway, but for now, we're as we were, just minus the dependency on Bootstrap (and with a couple of small improvements -- including preventing the guaranteed vertical overflow).