Upgrade SvgSaver
There are some small-ish fixes in a newer version than the one which has been vendored-in here.  Also, the version used here includes its own implementation (simply bundled, I assume) of FileSaver, so there's no need to be dragging that in from a CDN anymore (was there ever?).  Finally and most importantly, I'm using an unminified version, because I'm about to fork it and make some customizations.