Fix debugging template variable so it actually works
`site.debugging` is a boolean value, so comparing it to a string is not gonna fly.  I've also updated the unbundled includes so the correspond to the bundled ones.  I wonder if debugging is disabled for production builds?  If not, it might be better to reverse this check, and pass the variable to jekyll on the command line as desired.