Recalculate SVG padding
This should be considerably more robust.

Annotations (axis labels, mensuration symbols, and section names) are positioned off-canvas with respect to their own internal SVG elements.  In the browser, `overflow:visible` can be specified and this is not a problem, but in many other contexts (once the SVG has been exported), this setting is not respected and the annotations are clipped.  This commit processes the SVG before export to adjust `viewBox`s appropriately to ensure that nothing gets clipped.